# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Jabber Bot API
  description: 'XMPP Jabber bot with RESTful API for sending messages, monitoring status, and webhook integration. This API provides endpoints for sending XMPP messages to users and chat rooms, monitoring bot status and health, webhook management and statistics, and real-time message forwarding to configured webhooks. The bot connects to XMPP servers and provides a simple HTTP interface for bot operations.'
  version: 1.0.0
  contact:
    name: Jabber Bot API Support
    url: 'https://github.com/aleskapot/jabber-bot'
  license:
    name: MIT
servers:
  - url: 'http://localhost:8080'
    description: Development server
  - url: 'https://your-domain.com'
    description: Production server
paths:
...
  /api/v1/send:
    post:
      tags:
        - Messages
      summary: Send XMPP message
      description: |-
        Sends an XMPP message to a specified user JID. The message will be delivered immediately if the XMPP connection is active.

        **JID Format**: `username@domain.com/resource` or `username@domain.com`
        **Message Types**: 
        - `chat` (default) - One-to-one conversation
        - `groupchat` - Group chat message
        - `headline` - News headline
        - `normal` - Normal message

        **Authentication**: Requires API key via `API-Key` header or `Authorization: Bearer <key>`
      operationId: sendMessage
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            example:
              to: user@example.com
              body: 'Hello, world!'
              type: chat
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
              example:
                success: true
                message: Message sent successfully
                data:
                  to: user@example.com
                  type: chat
                  body_length: 13
                  sent_at: '2023-12-01T12:00:00Z'
                  request_id: abc123
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/send-muc:
    post:
      tags:
        - Messages
      summary: Send MUC message
      description: |-
        Sends a message to a Multi-User Chat (MUC) room. The message will be delivered to all room participants.

        **Room JID Format**: `roomname@conference.domain.com`
        **Subject**: Optional - sets/changes the room subject/topic

        **Authentication**: Requires API key via `API-Key` header or `Authorization: Bearer <key>`
      operationId: sendMUCMessage
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMUCMessageRequest'
            example:
              room: room@conference.example.com
              body: 'Hello, room!'
              subject: Room Topic
      responses:
        '200':
          description: MUC message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMUCMessageResponse'
              example:
                success: true
                message: MUC message sent successfully
                data:
                  room: room@conference.example.com
                  subject: Room Topic
                  body_length: 14
                  sent_at: '2023-12-01T12:00:00Z'
                  request_id: abc123
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/chat-state:
    post:
      tags:
        - Messages
      summary: Send Chat State Notification
      description: |-
        Sends a chat state notification (XEP-0085) to a specified user JID. Chat states indicate the typing status of a user in a chat conversation.

        **Chat States**:
        - `active` - User is actively participating in the chat session
        - `composing` - User is composing a message
        - `paused` - User has paused while composing a message
        - `inactive` - User has not been actively participating
        - `gone` - User has ended their participation in the chat session

        **JID Format**: `username@domain.com`

        **Authentication**: Requires API key via `API-Key` header or `Authorization: Bearer <key>`
      operationId: sendChatState
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendChatStateRequest'
            example:
              to: user@example.com
              state: composing
      responses:
        '200':
          description: Chat state sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendChatStateResponse'
              example:
                success: true
                message: Chat state sent successfully
                data:
                  to: user@example.com
                  state: composing
                  sent_at: '2023-12-01T12:00:00Z'
                  request_id: abc123
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/send-file:
    post:
      tags:
        - Messages
      summary: Send File via XMPP (XEP-0363)
      description: |-
        |-
                Uploads a file and sends it to a specified user JID via XMPP using XEP-0363 (HTTP File Upload) combined with XEP-0066 (Out-of-Band Data).

                **How it works:**
                1. File is uploaded to the server's storage directory
                2. File becomes accessible via HTTP URL: `GET /files/{filename}`
                3. XMPP message is sent with XEP-0066 OOB extension containing the HTTP URL
                4. Recipient's XMPP client can download the file from the provided URL

                **File Size Limit**: Configured via `file_transfer.max_size` (default: 10 MB)
                **Storage**: Files are saved to `file_transfer.storage_path` with unique names
                **Public URL**: Constructed using `file_transfer.base_url` (default: `http://localhost:8080/files`)
                **Retention**: Files are kept for `file_transfer.retain_days` (default: 7 days)

                **Supported Formats**: Any file type. MIME type is auto-detected from file extension.

                **Authentication**: Requires API key via `API-Key` header or `Authorization: Bearer <key>`
      operationId: sendFile
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                to:
                  type: string
                  description: Recipient JID
                  example: user@example.com
                description:
                  type: string
                  description: Optional description of the file
                  example: Project documentation
                file:
                  type: string
                  format: binary
                  description: File to upload
              required:
                - to
                - file
      responses:
        '200':
          description: File sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendFileResponse'
              example:
                success: true
                message: File sent successfully
                data:
                  to: user@example.com
                  description: Project documentation
                  file:
                    name: document.pdf
                    size: 1048576
                    type: application/pdf
                    url: 'http://localhost:8080/files/document_1700000000000.pdf'
                    uploaded_at: '2023-12-01T12:00:00Z'
                  sent_at: '2023-12-01T12:00:00Z'
                  request_id: abc123
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  schemas:
...
    SendMessageRequest:
      type: object
      required:
        - to
        - body
      properties:
        to:
          type: string
          description: JID of the message recipient
          format: xmpp-jid
          example: user@example.com
        body:
          type: string
          description: 'Message content (maximum 10,000 characters)'
          maxLength: 10000
          example: 'Hello, world!'
        type:
          type: string
          description: Message type
          enum:
            - chat
            - groupchat
            - headline
            - normal
          default: chat
          example: chat
    SendMUCMessageRequest:
      type: object
      required:
        - room
        - body
      properties:
        room:
          type: string
          description: JID of the MUC room
          format: xmpp-room-jid
          example: room@conference.example.com
        body:
          type: string
          description: 'Message content (maximum 10,000 characters)'
          maxLength: 10000
          example: 'Hello, room!'
        subject:
          type: string
          description: Optional room subject/topic
          maxLength: 200
          example: Room Topic
    SendChatStateRequest:
      type: object
      required:
        - to
        - state
      properties:
        to:
          type: string
          description: JID of the message recipient
          format: xmpp-jid
          example: user@example.com
        state:
          type: string
          description: Chat state notification
          enum:
            - active
            - composing
            - paused
            - inactive
            - gone
          example: composing
...
    SendChatStateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Chat state sent successfully
        data:
          $ref: '#/components/schemas/ChatStateResponseData'
...
    SendFileResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: File sent successfully
        data:
          type: object
          properties:
            to:
              type: string
              description: Recipient JID
              example: user@example.com
            description:
              type: string
              description: Optional description provided with the file
              example: Project documentation
            file:
              $ref: '#/components/schemas/FileInfo'
            sent_at:
              type: string
              format: date-time
              description: Timestamp when file was sent
              example: '2023-12-01T12:00:00Z'
            request_id:
              type: string
              description: Unique request identifier
              example: abc123
...
    SendMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Message sent successfully
        data:
          $ref: '#/components/schemas/MessageResponseData'
    SendMUCMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: MUC message sent successfully
        data:
          $ref: '#/components/schemas/MessageResponseData'
...
  responses:
    BadRequest:
      description: Bad request - validation error or invalid JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Invalid JID format
            code: 400
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized - valid API key required
            code: 401
    InternalServerError:
      description: Internal server error - XMPP errors or unexpected failures
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Failed to send XMPP message
            code: 500
    ServiceUnavailable:
      description: Service unavailable - XMPP connection lost
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: XMPP connection not available
            code: 503
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []
...